Move fuzzy finder tsx files to `client/web/src/components/fuzzyFinder`
The non-tsx files are still in the original `client/web/src/fuzzyFinder`
directory.
