+++ client/web/src/fuzzyFinder/CaseSensitiveFuzzySearch.test.ts
