+++ client/web/src/fuzzyFinder/Hasher.ts
